Safari-611.1.10/LayoutTests/js/dfg-dead-redundant-get-array-length-expected.txt

Tests that having a dead, and redundant, use of array.length doesn't cause compiler crashes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo([1, 2, 3]) is 3 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
